+++ src/server/entities/widgetData.ts
@@ -1,15 +1,13 @@
 import { z } from "zod";
-import {
-  calendarWidgetDataSchema,
-  linkWidgetDataSchema,
-  searchWidgetDataSchema,
-  timeWidgetDataSchema,
-  widgetTypeSchema,
-} from "~/utils/schema";
+import { widgetTypeSchema } from "~/utils/schema";
 import type { WidgetType } from "~/utils/types/widget";
+import { calendarWidgetDataSchema } from "../widgets/calendar/schema";
 import type { CalendarWidgetData } from "../widgets/calendar/types";
+import { linkWidgetDataSchema } from "../widgets/links/schema";
 import type { LinkWidgetData } from "../widgets/links/types";
+import { searchWidgetDataSchema } from "../widgets/search/schema";
 import type { SearchWidgetData } from "../widgets/search/types";
+import { timeWidgetDataSchema } from "../widgets/time/schema";
 import type { TimeWidgetData } from "../widgets/time/types";
 
 export class WidgetData {
